멤버들에게 정산을 요청하는 기능 API 작성

+++ src/main/java/com/core/backend/settlement/ui/dto/SettlementParticipantResponse.java
@@ -12,6 +12,6 @@ public class SettlementParticipantResponse {
 
     private Long id;
     private String participantName;
-    private int paymentAmount;
+    private Long paymentAmount;
     private boolean agreementStatus;
 }
